+++ Api/Data/SiteInterface.php
@@ -15,6 +15,10 @@ interface SiteInterface extends ExtensibleDataInterface
     const SITE_ID = 'site_id';
     const API_USERNAME = 'api_username';
     const API_PASSWORD = 'api_password';
+    const REPLACE_MEDIA_URLS = 'replace_media_urls';
+    const ENABLE_BLOG = 'enable_blog';
+    const BLOG_PREFIX = 'blog_prefix';
+    const BLOG_STORES = 'blog_stores';
 
     /**
      * Get site_id
@@ -106,4 +110,56 @@ interface SiteInterface extends ExtensibleDataInterface
      * @return SiteInterface
      */
     public function setApiPassword(?string $apiPassword);
+
+    /**
+    * Get replace_media_urls
+    * @return bool
+    */
+    public function getReplaceMediaUrls(): ?bool;
+
+    /**
+    * Set replace_media_urls
+    * @param bool $replaceMediaUrls
+    * @return SiteInterface
+    */
+    public function setReplaceMediaUrls(?bool $replaceMediaUrls);
+
+    /**
+     * Get enable_blog
+     * @return bool
+     */
+    public function getEnableBlog(): ?bool;
+
+    /**
+     * Set enable_blog
+     * @param bool $enableBlog
+     * @return SiteInterface
+     */
+    public function setEnableBlog(?bool $enableBlog);
+
+    /**
+     * Get blog_prefix
+     * @return string
+     */
+    public function getBlogPrefix(): ?string;
+
+    /**
+     * Set blog_prefix
+     * @param string $blogPrefix
+     * @return SiteInterface
+     */
+    public function setBlogPrefix(?string $blogPrefix);
+
+    /**
+     * Get blog_stores
+     * @return string
+     */
+    public function getBlogStores(): string;
+
+    /**
+     * Set blog_stores
+     * @param string $blogStores
+     * @return SiteInterface
+     */
+    public function setBlogStores(string $blogStores);
 }